requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
click==8.1.7
werkzeug==2.3.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
gunicorn==21.2.0
